+++ PetHealthCareSystem/src/Atomic Components/pages/LoginPage/LoginPage.scss
@@ -1,5 +1,3 @@
-$small-font-size: 0.813rem;
-
 .login {
   background-color: var(--LILY-WHITE);
   position: relative;
@@ -16,142 +14,7 @@ $small-font-size: 0.813rem;
     object-position: center;
   }
 
-  &__form {
-    position: relative;
-    border: 2px solid #FFFFFF;
-    margin-inline: 1.5rem;
-    padding: 2.5rem 1.5rem;
-    border-radius: 1rem;
-    background-color: #FFFFFF;
-  }
-
-  &__logo {
-    text-align: center;
-    margin-bottom: 20px;
-    &-img{
-      width: 120px;
-      height: 100px;
-      border-radius: 10px;
-    }
-  }
-
-  &__content,
-  &__box {
-    display: grid;
-  }
-
-  &__content {
-    row-gap: 1.75rem;
-    margin-bottom: 1.5rem;
-  }
-
-  &__box {
-    grid-template-columns: 1fr;
-    align-items: center;
-    border-bottom: 2px solid #000;
-  }
-
-  &__input {
-    width: 95%;
-    padding-block: 0.8rem;
-    background: none;
-    color: #000;
-    position: relative;
-    border: none;
-    outline: none;
-    z-index: 1;
-
-    &:focus + .login__label {
-      top: -12px;
-      font-size: $small-font-size;
-    }
-
-    &:not(:placeholder-shown).login__input:not(:focus) + .login__label {
-      top: -12px;
-      font-size: $small-font-size;
-    }
-  }
-
-  .password-toggle-btn {
-    background-color: transparent;
-    margin: 0;
-    padding: 0;
-    border: none;
-  }
-
-  &__box-input {
-    position: relative;
-  }
-
-  &__label {
-    position: absolute;
-    left: 0;
-    top: 13px;
-    transition: top 0.3s, font-size 0.3s;
-    color: #000;
-  }
-
-  &__check{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 1.5rem;
-  }
-
-  &__check-label {
-    color: #000;
-  }
-
-  &__check-label,
-  &__forgot,
-  &__register {
-    font-size: $small-font-size;
-  }
-
-  &__check-group {
-    column-gap: 0.5rem;
-  }
-
-  &__check-input {
-    width: 16px;
-    height: 16px;
-  }
-
-  &__forgot {
-    color: var(--JELLY-BEAN);
-    text-decoration: none;
-
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-
-  &__button {
-    width: 100%;
-    padding: 1rem;
-    border-radius: 0.5rem;
-    cursor: pointer;
-    margin-bottom: 2rem;
-  }
-
-  .register-line {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 2px;
-  }
-  
-  .login__register-link {
-    color: var(--JELLY-BEAN);
-    font-size: $small-font-size;
-    text-decoration: none;
-  
-    &:hover {
-      text-decoration: underline;
-    }
-  }  
-  
-  .copyright{
+  .copyright {
     position: absolute;
     bottom: 0;
     right: 0;
@@ -163,15 +26,5 @@ $small-font-size: 0.813rem;
 @media screen and (min-width: 576px) {
   .login {
     justify-content: center;
-
-    &__form {
-      width: 432px;
-      padding: 4rem 3rem 3.5rem;
-      border-radius: 1.5rem;
-    }
-
-    &__title {
-      font-size: 2rem;
-    }
   }
 }
