+++ PetHealthCareSystem/src/Atomic Components/templates/LoginPageTemplate/LoginPageTemplate.scss
@@ -1,0 +1,140 @@
+$small-font-size: 0.813rem;
+
+.login__form {
+  position: relative;
+  border: 2px solid #FFFFFF;
+  margin-inline: 1.5rem;
+  padding: 2.5rem 1.5rem;
+  border-radius: 1rem;
+  background-color: #FFFFFF;
+
+  @media screen and (min-width: 576px) {
+    width: 432px;
+    padding: 4rem 3rem 3.5rem;
+    border-radius: 1.5rem;
+  }
+}
+
+.login__logo {
+  text-align: center;
+  margin-bottom: 20px;
+
+  &-img {
+    width: 120px;
+    height: 100px;
+    border-radius: 10px;
+  }
+}
+
+.login__content {
+  display: grid;
+  row-gap: 1.75rem;
+  margin-bottom: 1.5rem;
+}
+
+.login__box {
+  display: grid;
+  grid-template-columns: 1fr;
+  align-items: center;
+  border-bottom: 2px solid #000;
+}
+
+.login__box-input {
+  position: relative;
+}
+
+.login__input {
+  width: 95%;
+  padding-block: 0.8rem;
+  background: none;
+  color: #000;
+  position: relative;
+  border: none;
+  outline: none;
+  z-index: 1;
+
+  &:focus + .login__label {
+    top: -12px;
+    font-size: $small-font-size;
+  }
+
+  &:not(:placeholder-shown).login__input:not(:focus) + .login__label {
+    top: -12px;
+    font-size: $small-font-size;
+  }
+}
+
+.login__label {
+  position: absolute;
+  left: 0;
+  top: 13px;
+  transition: top 0.3s, font-size 0.3s;
+  color: #000;
+}
+
+.password-toggle-btn {
+  background-color: transparent;
+  margin: 0;
+  padding: 0;
+  border: none;
+}
+
+.login__check {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 1.5rem;
+}
+
+.login__check-label {
+  color: #000;
+}
+
+.login__check-label,
+.login__forgot,
+.login__register {
+  font-size: $small-font-size;
+}
+
+.login__check-group {
+  column-gap: 0.5rem;
+}
+
+.login__check-input {
+  width: 16px;
+  height: 16px;
+}
+
+.login__forgot {
+  color: var(--JELLY-BEAN);
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.login__button {
+  width: 100%;
+  padding: 1rem;
+  border-radius: 0.5rem;
+  cursor: pointer;
+  margin-bottom: 2rem;
+}
+
+.register-line {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 2px;
+}
+
+.login__register-link {
+  color: var(--JELLY-BEAN);
+  font-size: $small-font-size;
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
